Ensure all swift files import UIKit if they need it.
We have been able to get away with not strictly adhering to this up until
now because we have been importing UIKit in the ObjC bridging header which
automatically makes it available to all swift files. That approach will
stop working once we remove the last of the ObjC files.